move APITimeout to days.py
